test mqtt 100.102 1880

+++ internal/dataProcessingService/server.go
@@ -65,9 +65,9 @@ func (config *Config) dataToWialonModule(dataChan, dataSourceChan, done chan str
 }
 
 func sendTimer(startTime time.Time, rate float64, speedCoef float64, startCourse int, courseDiffTrigger int) {
-	defFinishTime := startTime.Add(time.Duration(rate/(1.0+(getSpeedF64()*float64(speedCoef)*0.01))) * time.Second)
+	defFinishTime := startTime.Add(time.Duration(rate/(1.0+(getSpeedF64()*speedCoef*0.01))) * time.Second)
 	for defFinishTime.After(time.Now()) && diff(startCourse, getCourseInt()) < courseDiffTrigger {
-		time.Sleep(time.Duration(500/(1+getSpeedF64()*float64(speedCoef)*0.01)) * time.Millisecond)
+		time.Sleep(time.Duration(500/(1+getSpeedF64()*speedCoef*0.01)) * time.Millisecond)
 	}
 }
